Good news, this is actually working well enough to be used.  Bad news is it was a pretty hefty rewrite and I made no commits during it, sorry.
We now how functions VAPPLY and VFUNCALL which allow you to access data in a
normal way.  Use VMODF to modify versioned data.  The macros
WITH-VERSIONED-OBJECT(S) still exist but using them is a bit dangerous.

Performance is a bit worse than for a specialized data type like arrays.  I
think that I will implement a handful of specific data types in the future, but
right now the general interface works.

There are some benchmarking functions at the end of the file.  These will help
us in optimization.