Big changes to VMODF and RAISE-OBJECT! to allow for "replacement" version.
In these modifications, you replace the entire object with a new object.  This
is an important functionality because functional operations on an object may
produce another object that shares state with the object being versioned.  This
means that accessing other versions will screw with that new object with shared
state unless it also uses the locks and ensures it is current before accessing
it's data.

To insert a new object into the version tree, just use (vmodf v-obj new-object).
The fact that you specify an atom clues vmodf into the fact that you want to
make a "replacment" version.